requirements.txt
i should install those

streamlit
streamlit-chat
openai
python-dotenv